SC 110072: Separate batched components (part 3)
