+++ src/applications/appeals/995/components/EvidenceSummary.jsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState, useRef } from 'react';
 import { VaModal } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-
 import FormNavButtons from 'platform/forms-system/src/js/components/FormNavButtons';
 import { focusElement, scrollTo } from 'platform/utilities/ui';
 import { $ } from 'platform/forms-system/src/js/utilities/ui';
@@ -14,11 +13,9 @@ import {
 
 import { content } from '../content/evidenceSummary';
 
-import {
-  VaContent,
-  PrivateContent,
-  UploadContent,
-} from './EvidenceSummaryLists';
+import { EvidencePrivateContent } from './EvidencePrivateContent';
+import { EvidenceUploadContent } from './EvidenceUploadContent';
+import { EvidenceVaContent } from './EvidenceVaContent';
 
 import { EVIDENCE_LIMIT, LIMITATION_KEY, SC_NEW_FORM_DATA } from '../constants';
 import { customPageProps995 } from '../../shared/props';
@@ -221,15 +218,15 @@ const EvidenceSummary = ({
             {removeData?.name ? <strong>{` ${removeData.name}`}</strong> : null}
           </p>
         </VaModal>
-        <VaContent list={vaEvidence} {...props} />
-        <PrivateContent
+        <EvidenceVaContent list={vaEvidence} {...props} />
+        <EvidencePrivateContent
           list={privateEvidence}
           showLimitedConsentYN={showLimitedConsentYN}
           limitedConsent={limitedConsent}
           privacyAgreementAccepted={privacyAgreementAccepted}
           {...props}
         />
-        <UploadContent list={otherEvidence} {...props} />
+        <EvidenceUploadContent list={otherEvidence} {...props} />
 
         {content.addMoreLink()}
